+++ src/EtcdTerminal/Session/IConnectionSession.cs
@@ -1,10 +1,12 @@
 using EtcdTerminal.Configuration;
+using EtcdTerminal.Security;
 
 namespace EtcdTerminal.Session;
 
 public interface IConnectionSession
 {
 	EtcdConnectionConfig? Active { get; }
-	void Start(EtcdConnectionConfig config);
+	UserCapabilities Capabilities { get; }
+	void Start(EtcdConnectionConfig config, UserCapabilities capabilities);
 	void End();
 }